Allow LW in lists for DOM
LW can be used in lists for the day-of-month CRON field, for Lobaro CRONs at least.
Altered logic to this:

If 'L' is enabled and present, check there is no '/' (iterator) present,
or that the Lobaro + regular W option are enabled, 'LW' is present and a comma is there.

If one of the mentioned conditions is false, return an error.